spread methods by files
